+++ packages/okta-auth-js/webpack.common.config.js
@@ -8,7 +8,16 @@ module.exports = {
       { test: /\.json$/, loader: 'json' }
     ],
     rules: [
-      { test: /\.js$/, exclude: /node_modules/, loader: 'babel-loader' }
+      {
+        test: /\.js$/,
+        exclude: /node_modules/,
+        loader: 'babel-loader',
+        query: {
+          presets: ['@babel/env'],
+          plugins: ['@babel/plugin-transform-runtime'],
+          sourceType: 'unambiguous'
+        }
+      }
     ]
   },
   plugins: [